+++ src/beta/openai/chat.ts
@@ -3,7 +3,7 @@
 *   Copyright © 2025 NatML Inc. All Rights Reserved.
 */
 
-import type { PredictionService } from "../../services"
+import type { PredictionService, PredictorService } from "../../services"
 import type { RemotePredictionService } from "../remote"
 import { ChatCompletionsService } from "./completions"
 
@@ -15,9 +15,10 @@ export class ChatService {
     public readonly completions: ChatCompletionsService;
 
     public constructor(
+        predictors: PredictorService,
         predictions: PredictionService,
         remotePredictions: RemotePredictionService
     ) {
-        this.completions = new ChatCompletionsService(predictions, remotePredictions);
+        this.completions = new ChatCompletionsService(predictors, predictions, remotePredictions);
     }
 }