+++ src/beta/openai/openai.ts
@@ -3,7 +3,7 @@
 *   Copyright © 2025 NatML Inc. All Rights Reserved.
 */
 
-import type { PredictionService } from "../../services"
+import type { PredictorService, PredictionService } from "../../services"
 import type { RemotePredictionService } from "../remote"
 import { ChatService } from "./chat"
 import { EmbeddingsService } from "./embeddings"
@@ -21,10 +21,11 @@ export class OpenAIClient {
     public readonly embeddings: EmbeddingsService;
 
     public constructor(
+        predictors: PredictorService,
         predictions: PredictionService,
         remotePredictions: RemotePredictionService
     ) {
-        this.chat = new ChatService(predictions, remotePredictions);
-        this.embeddings = new EmbeddingsService(predictions, remotePredictions);
+        this.chat = new ChatService(predictors, predictions, remotePredictions);
+        this.embeddings = new EmbeddingsService(predictors, predictions, remotePredictions);
     }
 }