+++ src/types/dtype.ts
@@ -1,0 +1,31 @@
+/*
+*   Muna
+*   Copyright © 2025 NatML Inc. All Rights Reserved.
+*/
+
+/**
+ * Value data type.
+ */
+export type Dtype =
+    "bfloat16"  |
+    "float16"   |
+    "float32"   |
+    "float64"   |
+    "int8"      |
+    "int16"     |
+    "int32"     |
+    "int64"     |
+    "uint8"     |
+    "uint16"    |
+    "uint32"    |
+    "uint64"    |
+    "bool"      |
+    "string"    |
+    "list"      |
+    "dict"      |
+    "image"     |
+    "audio"     |
+    "video"     |
+    "model"     |
+    "binary"    |
+    "null";