+++ src/types/index.ts
@@ -3,6 +3,7 @@
 *   Copyright © 2025 NatML Inc. All Rights Reserved.
 */
 
+export * from "./dtype"
 export * from "./prediction"
 export * from "./predictor"
 export * from "./user"