+++ src/types/value.ts
@@ -4,33 +4,6 @@
 */
 
 /**
- * Value data type.
- */
-export type Dtype =
-    "bfloat16"  |
-    "float16"   |
-    "float32"   |
-    "float64"   |
-    "int8"      |
-    "int16"     |
-    "int32"     |
-    "int64"     |
-    "uint8"     |
-    "uint16"    |
-    "uint32"    |
-    "uint64"    |
-    "bool"      |
-    "string"    |
-    "list"      |
-    "dict"      |
-    "image"     |
-    "audio"     |
-    "video"     |
-    "model"     |
-    "binary"    |
-    "null";
-
-/**
  * Boolean typed array.
  */
 export class BoolArray extends Uint8Array {